Separate reading and writing of vcf into functions. Create sviper.h header file which does all the heavy lifting.
